+++ src/collatz_steps.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-uint_least64_t collatz_steps(uint_least64_t);
+double collatz_steps(double);
