+++ test/steps_test.c
@@ -3,8 +3,8 @@
 #include <collatz_steps.h>
 
 typedef struct {
-        uint_least64_t y;
-        uint_least64_t x;
+        double y;
+        double x;
 } test_case;
 
 const test_case cases[] = {
@@ -84,7 +84,7 @@ int main() {
         int failed = 0;
         clock_t start = clock();
         for(uint_least64_t i = 0; i < len(cases); i++) {
-                uint_least64_t got = collatz_steps(cases[i].x);
+                double got = collatz_steps(cases[i].x);
                 if(got != cases[i].y) {
                         printf("Failed test case: "
                                "{x: %llu, got: %llu, want: %llu}\n",
